Update Section Controller
