+++ src/server/controller/section/updateSection.ts
@@ -5,26 +5,27 @@ import { BaseProtocol, DB } from '@server/db'
 import { SectionRepository } from '@server/repository/assessment/section'
 import { ActivityLogRepository } from '@server/repository/public/activityLog'
 
-export const removeSection = async (
-  props: { assessment: Assessment; section: Section; user?: User },
+export const updateSection = async (
+  props: { user: User; assessment: Assessment; section: Section },
   client: BaseProtocol = DB
-): Promise<void> => {
-  const { assessment, section, user } = props
+): Promise<Section> => {
+  const { user, assessment, section } = props
 
   return client.tx(async (t) => {
-    const deletedSection = await SectionRepository.removeAssessmentSection({ assessment, section }, t)
+    const updatedSection = await SectionRepository.update({ section, assessment }, t)
 
     await ActivityLogRepository.insertActivityLog(
       {
         activityLog: {
-          target: { deletedSection },
+          target: updatedSection,
           section: 'section',
-          message: ActivityLogMessage.sectionDelete,
+          message: ActivityLogMessage.sectionUpdate,
           user,
         },
         assessment,
       },
       t
     )
+    return updatedSection
   })
 }
