+++ src/server/repository/assessment/section/index.ts
@@ -1,15 +1,15 @@
-import { createAssessmentSection } from './createSection'
+import { create } from './create'
 import { getMany } from './getMany'
 import { getManyMetadata } from './getManyMetadata'
 import { getOne } from './getOne'
-import { removeAssessmentSection } from './removeSection'
-import { updateAssessmentSection } from './updateSection'
+import { remove } from './remove'
+import { update } from './update'
 
 export const SectionRepository = {
   getMany,
   getManyMetadata,
   getOne,
-  createAssessmentSection,
-  removeAssessmentSection,
-  updateAssessmentSection,
+  create,
+  remove,
+  update,
 }
